Version 2.1
In this version was created the register page and its endpoints with its services and repositories functions. On the other hand in this version was created the home and login views, and the profile patient, profile therapist, and registerPatient views were created. Besides, the avatars view and elements were added to the system. Finally the profile patient view already is responsive, and the db was created just with USER, THERAPIST, PATIENT, and THERAPIST_USER (which represent relation between THERAPIST and PATIENT table).